+++ test_app/app/controllers/messages_controller.rb
@@ -1,6 +1,7 @@
 class MessagesController < ApplicationController
   def index
     @messages = Message.all
+    @dexie_stream = "message_channel"
   end
 
   def create